Refactor: smart and dump function
- create util files: pagination.js, post.js, search.js.
- move relative codes to new files
- let main flow explicit and clean.
