+++ js/utils/index.js
@@ -1,3 +1,6 @@
 // export all named functions in common.js file
 export * from './common';
 export * from './selectors';
+export * from './post';
+export * from './pagination';
+export * from './search';
