+++ js/utils/post.js
@@ -1,0 +1,59 @@
+import dayjs from "dayjs";
+import relativeTime from "dayjs/plugin/relativeTime";
+import { setTextContent, truncateText } from "./common";
+
+// to use function fromNow()
+dayjs.extend(relativeTime);
+
+export function createPostElement(post) {
+  if (!post) return;
+
+  try {
+    // find and clone template
+    const templateElement = document.getElementById('postItemTemplate')
+    if (!templateElement) return;
+    const liElement = templateElement.content.firstElementChild.cloneNode(true);
+    if (!liElement) return;
+    
+    setTextContent(liElement, '[data-id="title"]', post.title);
+    setTextContent(liElement, '[data-id="description"]', truncateText(post.description, 100));
+    setTextContent(liElement, '[data-id="author"]', post.author);
+
+    // calculate timespan
+    const timeSpan = dayjs(post.updatedAt).fromNow();
+    setTextContent(liElement, '[data-id="timeSpan"]', ` - ${timeSpan}`);
+
+    const thumbnailElement = liElement.querySelector('[data-id="thumbnail"]');
+    if (thumbnailElement) {
+      thumbnailElement.src = post.imageUrl;
+
+      thumbnailElement.addEventListener('error', () => {
+        console.log('load image error --> use default placeholder');
+        thumbnailElement.src = 'https://via.placeholder.com/1368x400?text=thumbnail';
+      });
+    }
+    
+    
+    // attach events
+
+    return liElement;
+  } catch (error) {
+    console.log('Failed to create post item' , error);
+  }
+}
+
+export function renderPostList(elementId, postList) {
+  console.log(postList);
+  if (!Array.isArray(postList)) return;
+
+  const ulElement = document.getElementById(elementId);
+  if (!ulElement) return;
+
+  // clear current list
+  ulElement.textContent = '';
+
+  postList.forEach(post => {
+    const liElement = createPostElement(post);
+    ulElement.appendChild(liElement);
+  });
+}
